add more logs

+++ keystone_spassword/contrib/spassword/backends/sql.py
@@ -62,8 +62,10 @@ class SPassword(Driver):
     def remove_user(self, user_id):
         session = sql.get_session()
         spassword_ref = session.query(SPasswordModel).get(user_id)
-        with session.begin():
-            session.delete(spassword_ref)
+        LOG.info('removing user %s from spassword' % user_id)
+        if spassword_ref:
+            with session.begin():
+                session.delete(spassword_ref)
 
     def set_user_creation_time(self, user):
         session = sql.get_session()
